+++ src/main/java/janggi/JanggiApplication.java
@@ -4,8 +4,8 @@ import janggi.controller.JanggiController;
 import janggi.database.DatabaseConnection;
 import janggi.repository.GameRepository;
 import janggi.repository.GimulRepository;
-import janggi.repository.dao.GameDao;
-import janggi.repository.dao.GimulDao;
+import janggi.repository.jdbc.JdbcGameRepository;
+import janggi.repository.jdbc.JdbcGimulRepository;
 import janggi.service.JanggiService;
 import janggi.view.InputView;
 import janggi.view.OutputView;
@@ -15,8 +15,8 @@ public class JanggiApplication {
     public static void main(String[] args) {
         DatabaseConnection.initSchema();
 
-        GameRepository gameRepository = new GameDao();
-        GimulRepository gimulRepository = new GimulDao();
+        GameRepository gameRepository = new JdbcGameRepository();
+        GimulRepository gimulRepository = new JdbcGimulRepository();
         JanggiService janggiService = new JanggiService(gameRepository, gimulRepository);
 
         JanggiController controller = new JanggiController(
